requirements.txt
pandas==2.3.3
numpy==2.3.5
scikit-learn==1.7.2
PyDrive2==1.21.3
gdown==5.2.0
PyYAML==6.0.3
pyarrow==22.0.0
matplotlib==3.10.7
requests==2.32.5
tqdm==4.67.1
